購入履歴の検索をPOSTに変更

+++ db/migrate/20130223102335_create_konyu_rirekis.rb
@@ -1,7 +1,7 @@
 class CreateKonyuRirekis < ActiveRecord::Migration
   def change
     create_table :konyu_rirekis do |t|
-      t.integer :konyuRirekiId
+      t.decimal :konyuRirekiId
       t.integer :kokyakuId
       t.integer :byoinCd
       t.string :ishiNm
